# src/Prdb.Fab.Host/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "prdb-fab",
    "description": "The API the browser side of prdb-fab talks to. Generated from the code at build time and committed: change an endpoint, not this file.",
    "version": "v1"
  },
  "paths": {
...
    "/api/releases": {
      "get": {
        "tags": [
          "Release discovery"
        ],
        "parameters": [
          {
            "name": "video",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "site",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "actor",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "state",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/IdentificationState"
            }
          },
          {
            "name": "indexer",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleasePage"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/releases/searches": {
      "post": {
        "tags": [
          "Release discovery"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManualSearchStartVerdict"
                }
              }
            }
          }
        }
      }
    },
    "/api/releases/searches/latest": {
      "get": {
        "tags": [
          "Release discovery"
        ],
        "parameters": [
          {
            "name": "video",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManualSearchView"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/releases/searches/{searchId}": {
      "get": {
        "tags": [
          "Release discovery"
        ],
        "parameters": [
          {
            "name": "searchId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManualSearchView"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/releases/searches/{searchId}/indexers/{indexerId}/retry": {
      "post": {
        "tags": [
          "Release discovery"
        ],
        "parameters": [
          {
            "name": "searchId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "indexerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManualSearchRetryVerdict"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/releases/discovery-routines/run-now": {
      "post": {
        "tags": [
          "Release discovery"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReleaseDiscoveryRunNowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseDiscoveryRunNowVerdict"
                }
              }
            }
          }
        }
      }
    },
    "/api/releases/{releaseId}/download/preview": {
      "post": {
        "tags": [
          "Release discovery"
        ],
        "parameters": [
          {
            "name": "releaseId",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadPreview"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/releases/{releaseId}/download": {
      "post": {
        "tags": [
          "Release discovery"
        ],
        "parameters": [
          {
            "name": "releaseId",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadVerdict"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DownloadCause": {
        "enum": [
          "Rejected",
          "Failed",
          "Unusable",
          "Vanished",
          "Abandoned",
          "Empty",
          null
        ]
      },
...
      "DownloadOutcome": {
        "enum": [
          "Pending",
          "Submitted",
          "Rejected",
          "SubmissionUnknown",
          "ConnectionProblem",
          "IndexerProblem",
          "ReleaseNotEligible",
          "NoReleasesLeft",
          "RetryBudgetSpent"
        ]
      },
...
      "DownloadPlanOutcome": {
        "enum": [
          "Ready",
          "ReleaseNotEligible",
          "NoReleasesLeft",
          "RetryBudgetSpent"
        ]
      },
      "DownloadPreview": {
        "required": [
          "outcome",
          "downloadId",
          "release",
          "downloadsSpent",
          "retryBudget",
          "detail"
        ],
        "type": "object",
        "properties": {
          "outcome": {
            "$ref": "#/components/schemas/DownloadPlanOutcome"
          },
          "downloadId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "release": {
            "$ref": "#/components/schemas/ReleaseChoice"
          },
          "downloadsSpent": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "retryBudget": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": "string"
          }
        }
      },
      "DownloadPreviewRequest": {
        "required": [
          "videoId"
        ],
        "type": "object",
        "properties": {
          "videoId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "DownloadRequest": {
        "required": [
          "downloadId",
          "videoId"
        ],
        "type": "object",
        "properties": {
          "downloadId": {
            "type": "string",
            "format": "uuid"
          },
          "videoId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "DownloadState": {
        "enum": [
          "Outstanding",
          "Completed",
          "Collected",
          "Failed",
          "Abandoned"
        ]
      },
      "DownloadVerdict": {
        "required": [
          "outcome",
          "downloadId",
          "state",
          "cause",
          "nzoId",
          "detail"
        ],
        "type": "object",
        "properties": {
          "outcome": {
            "$ref": "#/components/schemas/DownloadOutcome"
          },
          "downloadId": {
            "type": "string",
            "format": "uuid"
          },
          "state": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DownloadState"
              }
            ]
          },
          "cause": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DownloadCause"
              }
            ]
          },
          "nzoId": {
            "type": [
              "null",
              "string"
            ]
          },
          "detail": {
            "type": "string"
          }
        }
      },
...
      "IdentificationState": {
        "enum": [
          "Unexamined",
          "Unremarkable",
          "Awaiting",
          "Matched",
          "SiteOnly",
          "Ambiguous",
          "Unknown"
        ]
      },
...
      "ManualSearchIndexerView": {
        "required": [
          "indexerId",
          "indexer",
          "state",
          "startedAt",
          "finishedAt",
          "deferredUntil",
          "resultsSeen",
          "rowsAdded",
          "detail",
          "canRetry"
        ],
        "type": "object",
        "properties": {
          "indexerId": {
            "type": "string",
            "format": "uuid"
          },
          "indexer": {
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/ManualSearchIndexerState"
          },
          "startedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "finishedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "deferredUntil": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "resultsSeen": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "rowsAdded": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "canRetry": {
            "type": "boolean"
          }
        }
      },
      "ManualSearchPhase": {
        "enum": [
          "Queued",
          "Searching",
          "Deferred",
          "Identifying",
          "Complete",
          "Failed"
        ]
      },
      "ManualSearchRequest": {
        "required": [
          "videoId",
          "indexerId"
        ],
        "type": "object",
        "properties": {
          "videoId": {
            "type": "string",
            "format": "uuid"
          },
          "indexerId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          }
        }
      },
      "ManualSearchResultCounts": {
        "required": [
          "seen",
          "added",
          "pending",
          "awaiting",
          "matchedVideo",
          "matchedOtherVideo",
          "ambiguous",
          "siteOnly",
          "unknown",
          "unremarkable"
        ],
        "type": "object",
        "properties": {
          "seen": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "added": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pending": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "awaiting": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "matchedVideo": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "matchedOtherVideo": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "ambiguous": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "siteOnly": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "unknown": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "unremarkable": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "ManualSearchRetryOutcome": {
        "enum": [
          "Scheduled",
          "SearchNotFound",
          "IndexerNotSelected",
          "NotRetryable"
        ]
      },
      "ManualSearchRetryVerdict": {
        "required": [
          "outcome"
        ],
        "type": "object",
        "properties": {
          "outcome": {
            "$ref": "#/components/schemas/ManualSearchRetryOutcome"
          }
        }
      },
      "ManualSearchStartOutcome": {
        "enum": [
          "Started",
          "AlreadyRunning",
          "VideoNotFound",
          "TitleNotSearchable",
          "NoEnabledIndexers",
          "IndexerNotEnabled"
        ]
      },
      "ManualSearchStartVerdict": {
        "required": [
          "outcome",
          "searchId",
          "detail"
        ],
        "type": "object",
        "properties": {
          "outcome": {
            "$ref": "#/components/schemas/ManualSearchStartOutcome"
          },
          "searchId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "detail": {
            "type": "string"
          }
        }
      },
      "ManualSearchView": {
        "required": [
          "id",
          "videoId",
          "videoTitle",
          "query",
          "requestedAt",
          "phase",
          "active",
          "indexers",
          "results"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "videoId": {
            "type": "string",
            "format": "uuid"
          },
          "videoTitle": {
            "type": "string"
          },
          "query": {
            "type": "string"
          },
          "requestedAt": {
            "type": "string",
            "format": "date-time"
          },
          "phase": {
            "$ref": "#/components/schemas/ManualSearchPhase"
          },
          "active": {
            "type": "boolean"
          },
          "indexers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ManualSearchIndexerView"
            }
          },
          "results": {
            "$ref": "#/components/schemas/ManualSearchResultCounts"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ReleaseChoice": {
        "required": [
          "id",
          "indexerId",
          "indexerName",
          "derivedReleaseId",
          "title",
          "size",
          "confidence",
          "position",
          "exclusion"
        ],
        "type": "object",
        "properties": {
          "id": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "indexerId": {
            "type": "string",
            "format": "uuid"
          },
          "indexerName": {
            "type": "string"
          },
          "derivedReleaseId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "size": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "confidence": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/IdentificationConfidence"
              }
            ]
          },
          "position": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "exclusion": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ReleaseExclusion"
              }
            ]
          }
        }
      },
      "ReleaseContext": {
        "required": [
          "kind",
          "prdbId",
          "title"
        ],
        "type": "object",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/ReleaseContextKind"
          },
          "prdbId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "ReleaseContextKind": {
        "enum": [
          "Video",
          "Site",
          "Actor"
        ]
      },
...
      "ReleaseDiscoveryRoutineKind": {
        "enum": [
          "WantedSweep",
          "Screening",
          "BackwardsScreening",
          "Identification"
        ]
      },
      "ReleaseDiscoveryRunNowRequest": {
        "required": [
          "kind",
          "target"
        ],
        "type": "object",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/ReleaseDiscoveryRoutineKind"
          },
          "target": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          }
        }
      },
      "ReleaseDiscoveryRunNowVerdict": {
        "required": [
          "accepted",
          "detail"
        ],
        "type": "object",
        "properties": {
          "accepted": {
            "type": "boolean"
          },
          "detail": {
            "type": "string"
          }
        }
      },
...
      "ReleaseIndexer": {
        "required": [
          "id",
          "name",
          "rank"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "rank": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "ReleasePage": {
        "required": [
          "context",
          "releases",
          "indexers",
          "page",
          "pageSize",
          "total",
          "acquisition"
        ],
        "type": "object",
        "properties": {
          "context": {
            "$ref": "#/components/schemas/ReleaseContext"
          },
          "releases": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReleaseViewRow"
            }
          },
          "indexers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReleaseIndexer"
            }
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "total": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "acquisition": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/VideoAcquisition"
              }
            ]
          }
        }
      },
      "ReleaseViewRow": {
        "required": [
          "id",
          "title",
          "indexer",
          "size",
          "firstSeenAt",
          "identificationState",
          "confidence",
          "matchedBy",
          "video",
          "candidates",
          "siteOnlyMatch",
          "rankingPosition",
          "rankingExclusion",
          "applicableRules",
          "automaticDecisionReason"
        ],
        "type": "object",
        "properties": {
          "id": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "indexer": {
            "$ref": "#/components/schemas/ReleaseIndexer"
          },
          "size": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "firstSeenAt": {
            "type": "string",
            "format": "date-time"
          },
          "identificationState": {
            "$ref": "#/components/schemas/IdentificationState"
          },
          "confidence": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/IdentificationConfidence"
              }
            ]
          },
          "matchedBy": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/IdentificationRung"
              }
            ]
          },
          "video": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/IdentifiedVideo"
              }
            ]
          },
          "candidates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReleaseCandidate"
            }
          },
          "siteOnlyMatch": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SiteOnlyMatch"
              }
            ]
          },
          "rankingPosition": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "rankingExclusion": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ReleaseExclusion"
              }
            ]
          },
          "applicableRules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApplicableAutomationRule"
            }
          },
          "automaticDecisionReason": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AutomationDecisionReason"
              }
            ]
          }
        }
      },
...
      "VideoAcquisition": {
        "required": [
          "downloadsSpent",
          "retryBudget",
          "nextRelease",
          "downloads"
        ],
        "type": "object",
        "properties": {
          "downloadsSpent": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "retryBudget": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "nextRelease": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ReleaseChoice"
              }
            ]
          },
          "downloads": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DownloadSelectionRow"
            }
          }
        }
      },
...
    }
  },
...
}